pyproject.toml
[tool.poetry]
name = "lineage-bundle"
version = "0.0.1"
description = "Lineage generation bundle"
readme = "README.md"
repository = "https://github.com/daipe-ai/lineage-bundle"

license = "MIT"

authors = [
    "Jiri Koutny <jiri.koutny@datasentics.com>"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

include = ["*.yaml"]

packages = [
    { include = "lineagebundle", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyfony-bundles = "0.4.*"
console-bundle = "^0.4.*"
SQLAlchemy = "^1.3"
pyfony-sqlalchemy-bundle = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"
flake8-fixme = "^1.1.1"
pep8-naming = "^0.11.1"
poethepoet = "^0.10.0"
pyfony-core = "0.8.*"
datalake-bundle = "^0.6.1b5"
databricks-connect = "^7.3.6"

[tool.poetry.plugins."pyfony.bundle"]
create = "lineagebundle.LineageBundle:LineageBundle"

[pyfony.bootstrap]
container_init = "pyfonycore.container.container_init:init_with_current_bundle"
root_module_name = "lineagebundle"

[tool.black]
line-length = 140

[tool.poe.tasks]
black = "black src"
black-check = "black --check src"
flake8 = "flake8 --ignore=E501,W503 --exclude=src/lineagebundle/notebook/NotebookParserTest_sample src"
